requirements.txt
pillow>=10.0.0
pynput>=1.7.0
keyboard>=0.13.5
pyperclip>=1.8.0
